remove final headers from examples
