+++ example/gap_junctions/gap_junctions.cpp
@@ -10,6 +10,7 @@
 #include <nlohmann/json.hpp>
 
 #include <arbor/assert_macro.hpp>
+#include <arbor/cable_cell.hpp>
 #include <arbor/common_types.hpp>
 #include <arbor/context.hpp>
 #include <arbor/load_balance.hpp>
@@ -26,13 +27,20 @@
 
 #include <sup/ioutil.hpp>
 #include <sup/json_meter.hpp>
+#include <sup/json_params.hpp>
 
-#include "parameters.hpp"
+struct gap_params {
+    std::string name = "default";
+    unsigned n_cables = 3;
+    unsigned n_cells_per_cable = 5;
+    double stim_duration = 30;
+    double event_min_delay = 10;
+    double event_weight = 0.05;
+    double sim_duration = 100;
+    bool print_all = true;
+};
 
-#ifdef ARB_MPI_ENABLED
-#include <mpi.h>
-#include <arborenv/with_mpi.hpp>
-#endif
+gap_params read_options(int argc, char** argv);
 
 using arb::cell_gid_type;
 using arb::cell_lid_type;
@@ -371,3 +379,44 @@ arb::cable_cell gj_cell(cell_gid_type gid, unsigned ncell, double stim_duration)
     return cell;
 }
 
+gap_params read_options(int argc, char** argv) {
+    using sup::param_from_json;
+
+    gap_params params;
+    if (argc<2) {
+        std::cout << "Using default parameters.\n";
+        return params;
+    }
+    if (argc>2) {
+        throw std::runtime_error("More than command line one option not permitted.");
+    }
+
+    std::string fname = argv[1];
+    std::cout << "Loading parameters from file: " << fname << "\n";
+    std::ifstream f(fname);
+
+    if (!f.good()) {
+        throw std::runtime_error("Unable to open input parameter file: "+fname);
+    }
+
+    nlohmann::json json;
+    json << f;
+
+    param_from_json(params.name, "name", json);
+    param_from_json(params.n_cables, "n-cables", json);
+    param_from_json(params.n_cells_per_cable, "n-cells-per-cable", json);
+    param_from_json(params.stim_duration, "stim-duration", json);
+    param_from_json(params.event_min_delay, "event-min-delay", json);
+    param_from_json(params.event_weight, "event-weight", json);
+    param_from_json(params.sim_duration, "sim-duration", json);
+    param_from_json(params.print_all, "print-all", json);
+
+    if (!json.empty()) {
+        for (auto it=json.begin(); it!=json.end(); ++it) {
+            std::cout << "  Warning: unused input parameter: \"" << it.key() << "\"\n";
+        }
+        std::cout << "\n";
+    }
+
+    return params;
+}
